Fix crash on resume app from background & rename app theme
On resuming app from background state, ImageLoader which has initiated
previously may destroy. So that app may crash while tries to access it,
it is fixed by initiating the ImageLoader if an instance not exist.
And ReviewFragment will crash due to accessing destroyed activity in the
async task, it is fixed now.
Renamed app theme.
Modified travis to use sudo-enabled precise infrastructure and emulator.