Fix crash on resume app from background & rename app theme
On resuming app from background state, ReviewFragment will crash due to
accessing activity before it completely creates, it is fixed by
accessing activity on the onActivityCreated. And ImageLoader which has
initiated previously will destroy in the background state. So that app
may crash while it tries to access it, it is fixed by initiating the
ImageLoader if  an instance not exist.
Renamed app theme.