Provide API to start exam using the slug
An exam can be directly start using exam slug.

Result code sent to the onActivityResult() of activity which calls the
start exam API.(To support refresh & deep linking on user app).
* RESULT_OK - If the user started the exam(started the exam & pause or
completed).
* RESULT_CANCELLED - If the user didn't start the exam and navigated
back from the start exam screen. Additionally, an Intent is passed in
the result with a boolean based on the user action as below.
   * True - If the user pressed home icon(left navigation arrow) in the
            app.
   * False - If the user pressed back button on the device.

Provided sample implementation.