Possible fix for IndexOutOfBoundException
There are some random cases where section end is being called multiple
times. We need to fix this by refactoring the entire fragment properly.

As a result the currentSectionPosition index is being incremented
multiple times. During the last section, this value exceeds the maximum
number of sections.

For the time being we now get this value from the response received
from the API instead of blindly incrementing.