Fix exam ending on second last section
When second last section is ended, current section position will become
last section position. In earlier code to end exam, we check if current
section is the last section which is wrong as current section will always
be last section and its state is Not Started

So to end exam we check if current section is last section and if it is completed.