Make track selection dialog reusable for both player and download (#241)
### Changes done
Right now track selection dialog box can be used only by the player, but for downloading also we need track selection dialog box

The Download APIs of ExoPlayer won't expose TrackSelection handle. They only give the parameters and mappedTrackInfo

So instead of passing trackselector we can pass only parameters and mappedTrackinfo to Track Selection Dialog.
 
In addition to this, the On click of track selection box is changed as an interface. This is done so that upper layer can implement it as per requirements.