Fixed NetworkBoundResource degrade
We were using transformations map to avoid using MediatorLiveData.

However Transformations map was internally using MediatorLiveData.
MediatorLiveData expects to be observed for the onChange to be called.

Since we just need a way to observe once the room db, we accomplish the
same via an extension.

Also fixed test case where were not verifying but expecting no more
interaction.