fix: Add support to open Download activity (#548)
- In this commit #32ebd452, we implemented download support for attachments,
and once the download is complete, a notification is displayed. We also
added a pending intent to that notification. However, we encountered an
issue where the pending intent only works for opening PDF files, while
other file types cannot be opened.
- To address this problem, we made a commit that handles the
differentiation between PDF files, image files, and other file types. If
the downloaded file is a PDF, we open the PDF activity. If the
downloaded file is an image, we open the image viewer activity.
Otherwise, if it is any other file type, we open the download activity.
This allows for the appropriate handling of different file types in the
application.