fix: Disable onClick event for helpline images in quiz exam (#606)
- In the context of exam templates, we employ a Webview. We create HTML
data and feed it to the Webview. If the Webview includes any image tags,
we attach a shared listener to open images in a new window. However, for
helpline images within the quiz template, we wish to avoid displaying
them in a new window, so we exclude the process of adding the onclick
event for these specific images.