fix: Unable to upload a file in file type questions. (#610)
- When a user tries to upload a file in response to a question about
file types, we check if the user has given permission for storage
access. If permission is granted, we proceed to open the file manager.
However, if permission is not granted, we prompt the user to provide the
necessary permission.
- If the user has previously denied permission, we cannot request it
again. In this commit, we have included the feature to display a message
in a snackbar, notifying the user that the required permission is
essential if they have previously denied it.