fix: Prevent users from copying a YouTube video link while in fullscreen. (#613)
- In commit b48068f, we deactivated the long-press option. However, when
the user transitions to fullscreen mode, we alter the view so the
disabled long-press option is ineffective. In this particular commit, we
addressed this issue by disabling the long-press option when switching
to fullscreen mode. This is achieved through the use of the
disableLongPress variable, which is set to its default value of false.