enhancement: Add a question report button on the custom test solution page (#623)
- In commit 489ee14, we removed the functionality to report questions
for the custom test solution. This change was necessary because it
required an examId that we don't have for custom tests.
- Recently, we added support for reporting questions on the web, and now
we can report questions without needing an examId. As a result, in this
commit and onwards, we permanently display the 'Report Question' button
on the solution page for both normal exams and custom tests.
- We do not include the `examId` as param when reporting a question for
a custom test.